git-svn-id: file:///svn/phpbb/trunk@9404 89ea8834-ac86-4346-8a33-228a782c2dd0

+++ phpBB/modules/acp/acp_users.php
@@ -971,6 +971,7 @@ class acp_users
 					{
 						$sql = 'DELETE FROM ' . LOG_TABLE . '
 							WHERE log_type = ' . LOG_USERS . "
+							AND reportee_id = $user_id
 							$where_sql";
 						phpbb::$db->sql_query($sql);
